refactor(solver): isolate compound simplification hot loop (#14351)
